requirements.txt
PyYAML==5.3.1
pylint #анализатор кода https://python-scripts.com/code-analysis
